refactor(matrices): introduce RepMatrix class
