+++ pytest.ini
@@ -22,7 +22,5 @@ norecursedirs =
     sympy/parsing/autolev/test-examples
     sympy/integrals/rubi/rubi_tests
 
-doctestplus = enabled
-
 markers =
     nocache_fail
